requirements.txt
amqp==2.6.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
async-timeout==4.0.3
backcall==0.2.0
billiard==3.6.4.0
boto3==1.28.49
botocore==1.31.49
celery==4.4.7
celery-redbeat==2.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.1
django-cors-headers==4.2.0
django-environ==0.11.2
django-phonenumber-field==7.1.0
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
exceptiongroup==1.1.3
executing==1.2.0
google-auth==2.25.2
gunicorn==21.2.0
hashids==1.3.1
httpie==3.2.2
idna==3.4
ipython==8.15.0
jedi==0.19.0
jmespath==1.0.1
kombu==4.6.11
Markdown==3.4.4
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
packaging==23.1
parso==0.8.3
pexpect==4.8.0
phonenumberslite==8.13.20
pickleshare==0.7.5
Pillow==10.0.1
prompt-toolkit==3.0.39
psycopg2==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycountry==22.3.5
Pygments==2.16.1
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.2
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tenacity==8.2.3
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==1.3.0
wcwidth==0.2.6
